STERI016-69-flip-cards-am: Remove tiles and apply the changes in the flip-cards block
