--- blocks/tiles/tiles.css
@@ -1,134 +1,0 @@
-.tiles {
-  display: grid;
-  gap: 43px;
-  margin: 4px 0;
-  font-size: var(--body-font-size-s);
-}
-
-.tiles a {
-  text-decoration: none;
-}
-
-.tiles img {
-  width: 100%;
-  filter: invert(0%) sepia(0%) saturate(0%) hue-rotate(0deg) brightness(0%) contrast(100%);
-}
-
-.tiles.gray img {
-  filter: invert(0%) sepia(0%) saturate(0%) hue-rotate(0deg) brightness(0%) contrast(100%);
-}
-
-.tiles.blue img{
-  filter: none;
-}
-
-.tiles > a > div > div:first-child .icon {
-  width: 60px;
-  height: auto;
-}
-
-
-.tiles > a > div {
-  background: var(--color-grey200);
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  justify-content: space-between;
-  padding: 32px;
-  border-radius: 12px;
-}
-
-.tiles.gray > a > div {
-  background: var(--color-grey200);
-}
-
-.tiles.blue > a > div {
-  background: var(--color-bg-tertiary);
-}
-
-.tiles > a > div > div:last-of-type {
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  gap: 8px;
-}
-
-.tiles > a > div .clamp-title {
-  display: -webkit-box;
-  -webkit-line-clamp: 2;
-  -webkit-box-orient: vertical;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  margin-bottom: 0;
-}
-
-.tiles > a > div .clamp-description {
-  display: -webkit-box;
-  -webkit-line-clamp: 5;
-  -webkit-box-orient: vertical;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-
-.tiles.blue > a > div .clamp-title, .tiles.blue > a > div .clamp-description {
-  color: var(--color-text-tertiary);
-}
-
-.tiles > a > div a.button.primary {
-  margin: 0;
-  padding: 8px 24px;
-  line-height: 160%;
-  letter-spacing: 0.32px;
-}
-
-.tiles > a > div a.button.primary .icon {
-  --icon-color: var(--color-black);
-
-  display: inline-block;
-  margin: 0 auto;
-  height: 24px;
-  width: 24px;
-  background-color: var(--icon-color);
-  mask-repeat: no-repeat;
-  mask-position: center center;
-  mask-size: contain;
-  mask-image: var(--mask-image);
-}
-
-@media (min-width: 992px /* breakpoint-l */) {
-  .tiles {
-    display: flex;
-    gap: 8px;
-  }
-
-  .tiles > a {
-    flex: 1;
-  }
-
-  .tiles > a > div {
-    height: 420px;
-  }
-
-  .tiles.tiles-4-cols > a > div {
-    height: 472px;
-  }
-}
-
-/* stylelint-disable-next-line no-descending-specificity */
-main .section.tiles-container {
-  padding: 32px 0;
-}
-
-main .section.tiles-container.gray-background {
-  padding-top: 0;
-}
-
-/* stylelint-disable-next-line no-descending-specificity */
-main .section.tiles-container > .tiles-wrapper {
-  padding: 0 32px;
-}
-
-main .section.tiles-container > div.tiles-wrapper {
-  margin-bottom: 0;
-  margin-top: 0;
-}
